Rework threading code and Test/TLS improvements (#14)
Reworked the threading code the debuggee and debug adapter workers to
remove some of the extra locks and complications involved. This has
resulted in code that is covered by tests a bit more and should
hopefully limit the chances of a deadlock.

Parse of this change is adding support for different socket types like a
Unix Domain Socket or just a standard TCP socket. The default socket
type used is a UDS but a TCP socket can be requested in the
attach/listen scenario to still allow remote debugging. A UDS is the
default because it will only allow local communication and is restricted
to the current user only.

Added support for TLS client certificate authentication when creating an
ansibug listener. When enabled, only clients who provide a certificate
signed by a CA in the bundle provided to the listener will be allowed to
connect.

Improves the socket code to cover more use cases and fixes up some minor
problems encountered during testing of the new scenarios.