Implement terminate vs disconnect functionality
Adds support for either disconnecting or terminating a debug session. By
default a client typically will use the terminate/stop request for
a launch session and a disconnect request for an attach session. This
implements the handlers for both these types in Ansible and ensures the
session is still ended cleanly when either is requested.